Adjust how Stitch is checked out and make more installs conditional

+++ Realm/ObjectServerTests/setup_baas.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 ######################################
 #
@@ -18,62 +18,57 @@
 #
 ######################################
 
+set -e
+set -o pipefail
+
+source_root="$(cd "$(dirname "$0")"/../..; pwd)"
+export $(xargs < "${source_root}/dependencies.list")
+
 mongodb_version=4.2.5
-go_version=1.13
+go_version=1.14
 node_version=8.11.2
-stitch_rev=ce699769dbbb51162bf71567f95aa3187814685e
 
 mongodb_url="https://fastdl.mongodb.org/osx/mongodb-macos-x86_64-$mongodb_version.tgz"
 transpiler_target="node8-macos"
 server_stitch_lib_url="https://s3.amazonaws.com/mciuploads/mongodb-mongo-master/stitch-support/macos-debug/e791a2ea966bb302ff180dd4538d87c078e74747/stitch-support-4.3.2-721-ge791a2e-patch-5e2a6ad2a4cf473ae2e67b09.tgz"
+mongodb_directory="mongodb-macos-x86_64-$mongodb_version"
 
-BASE_DIR=`pwd`
-BUILD_DIR=`pwd`/build/baas
+BASE_DIR=$(pwd)
+BUILD_DIR=${BASE_DIR}/build/baas
 PID_FILE="$BUILD_DIR/pid.txt"
 
-function setup_mongod() {
-    set -e
-    if [ ! -d "mongodb-macos-x86_64-$mongodb_version" ]
-    then
+setup_mongod() {
+    if [ ! -d "$mongodb_directory" ]; then
         curl --silent ${mongodb_url} | tar xz
-        pushd mongodb-*
-        mkdir db_files
-        popd
+        mkdir "$mongodb_directory/db_files"
     fi
 }
 
-function run_mongod() {
-    set -e
-    pushd mongodb-*
+run_mongod() {
     echo "starting mongod..."
-    ./bin/mongod --dbpath ./db_files --port 26000 --replSet test --pidfilepath $PID_FILE
-    popd
+    "$mongodb_directory/bin/mongod" \
+        --dbpath "$mongodb_directory/db_files" \
+        --bind_ip 127.0.0.1 \
+        --port 26000 \
+        --replSet test \
+        --pidfilepath "$PID_FILE"
 }
 
-function wait_for_mongod() {
-    set -e
-    pushd mongodb-*
+wait_for_mongod() {
     echo "waiting for mongod to start up"
-    ./bin/mongo --nodb --eval 'assert.soon(function(x){try{var d = new Mongo("localhost:26000"); return true}catch(e){return false}}, "timed out connecting")'
-    ./bin/mongo --port 26000 --eval 'rs.initiate()'
+    "$mongodb_directory/bin/mongo" --nodb --eval 'assert.soon(function(x){try{var d = new Mongo("127.0.0.1:26000"); return true}catch(e){return false}}, "timed out connecting")'
+    "$mongodb_directory/bin/mongo" --port 26000 --eval 'rs.initiate()'
     echo "mongod is up."
-    popd
 }
 
-function shutdown_mongod() {
-    set -e
-    pushd mongodb-*
-    ./bin/mongo --port 26000 admin --eval "db.adminCommand({replSetStepDown: 0, secondaryCatchUpPeriodSecs: 0, force: true})"
-    ./bin/mongo --port 26000 admin --eval "db.shutdownServer()"
+shutdown_mongod() {
+    "$mongodb_directory/bin/mongo" --port 26000 admin --eval "db.adminCommand({replSetStepDown: 0, secondaryCatchUpPeriodSecs: 0, force: true})"
+    "$mongodb_directory/bin/mongo" --port 26000 admin --eval "db.shutdownServer()"
     echo "mongod is down."
-    popd
 }
 
-function setup_stitch() {
-    set -e
-    OG_DIR=`pwd`
-
-    cd $BASE_DIR
+setup_stitch() {
+    pushd "$BASE_DIR"
 
     echo "setting up stitch"
 
@@ -84,62 +79,66 @@ function setup_stitch() {
     echo "checking out stitch"
 
     cd stitch
-    
-    git checkout $stitch_rev
 
-    echo "downloading mdb dylibs"
-    mkdir -p etc/dylib
-    cd etc/dylib
-    curl -s "https://s3.amazonaws.com/mciuploads/mongodb-mongo-master/stitch-support/macos-debug/e791a2ea966bb302ff180dd4538d87c078e74747/stitch-support-4.3.2-721-ge791a2e-patch-5e2a6ad2a4cf473ae2e67b09.tgz" | tar xvfz - --strip-components=1
-    cd $BASE_DIR
+    if [ -d .git ]; then
+        git checkout "$STITCH_VERSION"
+    fi
 
-    echo "downloading update_doc"
-    curl --silent -O "https://s3.amazonaws.com/stitch-artifacts/stitch-mongo-libs/stitch_mongo_libs_osx_patch_cbcbfd8ebefcca439ff2e4d99b022aedb0d61041_59e2b7a5c9ec4432c400181c_17_10_15_01_19_33/update_doc"
-    echo "downloading assisted_agg"
-    curl --silent -O "https://s3.amazonaws.com/stitch-artifacts/stitch-mongo-libs/stitch_mongo_libs_osx_patch_cbcbfd8ebefcca439ff2e4d99b022aedb0d61041_59e2b7ab2a60ed5647001827_17_10_15_01_19_39/assisted_agg"
-    chmod +x update_doc
-    chmod +x assisted_agg
+    if [ ! -d etc/dylib ]; then
+        echo "downloading mdb dylibs"
+        mkdir -p etc/dylib
+        curl -s "${server_stitch_lib_url}" \
+            | tar xvfz - --strip-components=1 -C etc/dylib
+    fi
 
-    mkdir -p stitch/etc/dylib
-    cd stitch/etc/dylib
-    curl -s ${server_stitch_lib_url} | tar xvfz - --strip-components=1
-    cd ../../../
+    if [ ! -f update_doc ]; then
+        echo "downloading update_doc"
+        curl --silent -O "https://s3.amazonaws.com/stitch-artifacts/stitch-mongo-libs/stitch_mongo_libs_osx_patch_cbcbfd8ebefcca439ff2e4d99b022aedb0d61041_59e2b7a5c9ec4432c400181c_17_10_15_01_19_33/update_doc"
+        chmod +x update_doc
+    fi
+    if [ ! -f assisted_agg ]; then
+        echo "downloading assisted_agg"
+        curl --silent -O "https://s3.amazonaws.com/stitch-artifacts/stitch-mongo-libs/stitch_mongo_libs_osx_patch_cbcbfd8ebefcca439ff2e4d99b022aedb0d61041_59e2b7ab2a60ed5647001827_17_10_15_01_19_39/assisted_agg"
+        chmod +x assisted_agg
+    fi
 
-    echo "building transpiler"
-    cd stitch/etc/transpiler
-    if !command -v node >/dev/null 2>&1; then
+    if ! command -v node >/dev/null 2>&1; then
         echo "downloading node 🚀"
         curl -O "https://nodejs.org/dist/v8.11.2/node-v$node_version-darwin-x64.tar.gz"
         tar zxf node-v8.11.2-darwin-x64.tar.gz
-        export PATH=`pwd`/node-v8.11.2-darwin-x64/bin/:$PATH
+        export PATH="$(pwd)/node-v8.11.2-darwin-x64/bin/:$PATH"
     fi
 
-    rm -rf $HOME/.yarn
-    curl -o- -L https://yarnpkg.com/install.sh | bash
-    export PATH="$HOME/.yarn/bin:$HOME/.config/yarn/global/node_modules/.bin:$PATH"
-    yarn install && yarn run build -t ${transpiler_target}
+    if ! command -v yarn >/dev/null 2>&1; then
+        rm -rf "$HOME/.yarn"
+        curl -o- -L https://yarnpkg.com/install.sh | bash
+        export PATH="$HOME/.yarn/bin:$HOME/.config/yarn/global/node_modules/.bin:$PATH"
+    fi
 
-    cd $BASE_DIR
-    export ROOT_DIR=`pwd`
-    export PATH=$ROOT_DIR/:$PATH
+    echo "building transpiler"
+    cd etc/transpiler
+    yarn install && yarn run build -t "${transpiler_target}"
 
-    echo "downloading go"
-    curl --silent "https://dl.google.com/go/go$go_version.darwin-amd64.tar.gz" | tar xz
-    export GOROOT=$ROOT_DIR/go
-    export PATH=$GOROOT/bin:$PATH
+    cd "$BASE_DIR"
+    if ! command -v go >/dev/null 2>&1; then
+        echo "downloading go"
+        curl --silent "https://dl.google.com/go/go$go_version.darwin-amd64.tar.gz" | tar xz
+        export GOROOT="$BASE_DIR/go"
+        export PATH="$GOROOT/bin:$PATH"
+    fi
 
-    export STITCH_PATH=$ROOT_DIR/stitch
+    export STITCH_PATH="$BASE_DIR/stitch"
     export PATH="$PATH:$STITCH_PATH/etc/transpiler/bin"
     export LD_LIBRARY_PATH="$STITCH_PATH/etc/dylib/lib"
     echo "running stitch"
-    cd $STITCH_PATH
+    cd "$STITCH_PATH"
     go run cmd/auth/user.go addUser \
         -domainID 000000000000000000000000 \
-        -mongoURI mongodb://localhost:26000 \
+        -mongoURI mongodb://127.0.0.1:26000 \
         -salt 'DQOWene1723baqD!_@#' \
         -id "unique_user@domain.com" \
         -password "password"
-    cd $OG_DIR
+    popd
 }
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -147,11 +146,12 @@ function setup_stitch() {
 
 echo "Running with ACTION=${ACTION}"
 
-function build_action() {
+build_action() {
     echo "building baas"
-    if [ ! -d "mongodb-macos-x86_64-$mongodb_version" ]; then
-        mkdir -p $BUILD_DIR
-        cd $BUILD_DIR
+    trap shutdown_mongod ERR
+    if [ ! -d "${BUILD_DIR}/mongodb-macos-x86_64-$mongodb_version" ]; then
+        mkdir -p "$BUILD_DIR"
+        cd "$BUILD_DIR"
         setup_mongod
         run_mongod &
         wait_for_mongod
@@ -162,10 +162,10 @@ function build_action() {
     fi
 }
 
-function clean_action() {
+clean_action() {
     echo "cleaning baas"
     shutdown_mongod
-    rm -rf $BUILD_DIR
+    rm -rf "$BUILD_DIR"
 }
 
 case $1 in
